move components to folder
